chore(*): move shared search sub-components into a folder
