+++ frontend/src/pages/np/components/search/SearchBar.tsx
