+++ frontend/src/pages/np/components/search/TemplateFilterSidebar.tsx
